+++ Projects/Profile/Sources/Views/Components/ProfileRow.swift
@@ -35,8 +35,6 @@ struct ProfileRow: View {
         switch row {
         case .myInfo:
             router.push(ProfileRouterType.myInfo)
-        case .appInfo:
-            router.push(ProfileRouterType.appInfo)
         case .settings:
             router.push(ProfileRouterType.settings)
         case .eurobonus:
@@ -49,8 +47,8 @@ struct ProfileRow: View {
             router.push(ProfileRouterType.certificates)
         case .claimHistory:
             router.push(ProfileRouterType.claimHistory)
-        case .legal:
-            router.push(ProfileRouterType.legal)
+        case .information:
+            router.push(ProfileRouterType.information)
         }
     }
 }
